requirements.txt
Django == 3.2.5
Markdown == 3.3.4
PyMySQL == 1.0.2
asgiref == 3.4.1
django-filter == 2.4.0
djangorestframework == 3.12.4
mysqlclient == 2.0.3
pip == 21.1.3
pytz == 2021.1
setuptools == 57.4.0
sqlparse == 0.4.1
pillow == 8.3.1